refact : 맵 카테고리바, 바텀시트 리팩토링
